DO THE THINGS! TEST THE STUFF!
This is a modified matrix that tests the changes in the following PRs:

- https://github.com/ManageIQ/manageiq/pull/19547
- https://github.com/ManageIQ/manageiq-gems-pending/pull/454

but replaces ManageIQ/manageiq#19547 with the following branch:

https://github.com/NickLaMuro/manageiq/tree/rebase_testing

Which does a few extra steps to attempt to graft commits prior to
rebasing.

Unsure of the `git` results, but before open a separate PR, this should
test to see if specs still pass.