... in which the hero starts to lose hope...
"Gripping to what remnants of his sanity remain, our hero uses 'bash' as
his last resort to slay the CI build and complete his quest.

In the process, a 'man bash' was required, which briefly staggered and
disoriented the hero as he attempted to decipher the cryptic writings
left by scholars of old, in a last ditch effort to perform the proper
incantations necessary.

As he committed to his endeavour, only time will tell if the beast can
truly be slain with this effort... but the future seems bleak, because
if this fails... well..."